feat: add policies

+++ app/Http/Controllers/ItemController.php
@@ -11,11 +11,10 @@ use Inertia\Response;
 
 class ItemController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index(): Response
     {
+        $this->authorize('viewAny', Item::class);
+
         return Inertia::render('cashier/items/index', [
             'items' => Item::all(),
         ]);
@@ -26,14 +25,13 @@ class ItemController extends Controller
      */
     public function create(): Response
     {
+        $this->authorize('create', Item::class);
         return Inertia::render('cashier/items/create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(StoreItemRequest $request): RedirectResponse
     {
+        $this->authorize('create', Item::class);
         $item = new Item();
         $item->name = $request->validated('name');
         $item->price = $request->validated('price');
@@ -43,31 +41,26 @@ class ItemController extends Controller
         return redirect()->route('cashier.items.index')->with('success', 'Item created successfully.');
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(Item $item): Response
     {
+        $this->authorize('view', $item);
         return Inertia::render('cashier/items/show', [
             'item' => $item,
         ]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit(Item $item): Response
     {
+        $this->authorize('update', $item);
         return Inertia::render('cashier/items/edit', [
             'item' => $item,
         ]);
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(UpdateItemRequest $request, Item $item): RedirectResponse
     {
+        $this->authorize('update', $item);
+
         $item = Item::findOrFail($item->id);
         $item->name = $request->validated('name');
         $item->price = $request->validated('price');
@@ -77,11 +70,10 @@ class ItemController extends Controller
         return redirect()->route('cashier.items.index')->with('success', 'Item updated successfully.');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Item $item): RedirectResponse
     {
+        $this->authorize('delete', $item);
+
         $item->delete();
 
         return redirect()->route('cashier.items.index')->with('success', 'Item deleted successfully.');
